# docs/contracts/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Bookstore Microservices API",
    "description": "Complete API specification for the Bookstore microservices platform including Catalog, Order, and User services accessible through the API Gateway",
    "version": "1.0.0",
    "contact": {
      "name": "Bookstore API Support"
    }
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "API Gateway - Local Development"
    }
  ],
...
  "paths": {
    "/api/v1/books": {
      "get": {
        "tags": ["Books"],
        "summary": "Get all books",
        "description": "Retrieves all books from the catalog",
        "operationId": "getAllBooks",
        "responses": {
          "200": {
            "description": "Successfully retrieved all books",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/BookDto"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Books"],
        "summary": "Create a new book",
        "description": "Adds a new book to the catalog",
        "operationId": "createBook",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BookDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Book created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BookDto"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input or ISBN already exists",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/books/{id}": {
      "get": {
        "tags": ["Books"],
        "summary": "Get book by ID",
        "description": "Retrieves a specific book by its unique identifier",
        "operationId": "getBookById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "ID of the book to retrieve",
            "schema": {
              "type": "integer",
              "format": "int64",
              "minimum": 1
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Book found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BookDto"
                }
              }
            }
          },
          "404": {
            "description": "Book not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": ["Books"],
        "summary": "Update a book",
        "description": "Updates an existing book's information",
        "operationId": "updateBook",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "ID of the book to update",
            "schema": {
              "type": "integer",
              "format": "int64",
              "minimum": 1
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BookDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Book updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BookDto"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Book not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": ["Books"],
        "summary": "Delete a book",
        "description": "Removes a book from the catalog",
        "operationId": "deleteBook",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "ID of the book to delete",
            "schema": {
              "type": "integer",
              "format": "int64",
              "minimum": 1
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Book deleted successfully"
          },
          "404": {
            "description": "Book not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/books/isbn/{isbn}": {
      "get": {
        "tags": ["Books"],
        "summary": "Get book by ISBN",
        "description": "Retrieves a book by its ISBN",
        "operationId": "getBookByIsbn",
        "parameters": [
          {
            "name": "isbn",
            "in": "path",
            "required": true,
            "description": "ISBN of the book to retrieve",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Book found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BookDto"
                }
              }
            }
          },
          "404": {
            "description": "Book not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/books/search": {
      "get": {
        "tags": ["Books"],
        "summary": "Search books",
        "description": "Searches books by title or author",
        "operationId": "searchBooks",
        "parameters": [
          {
            "name": "searchTerm",
            "in": "query",
            "required": true,
            "description": "Search term to match against title or author",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Search completed successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/BookDto"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/books/category/{category}": {
      "get": {
        "tags": ["Books"],
        "summary": "Get books by category",
        "description": "Retrieves all books in a specific category",
        "operationId": "getBooksByCategory",
        "parameters": [
          {
            "name": "category",
            "in": "path",
            "required": true,
            "description": "Category/genre to filter by",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully retrieved books",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/BookDto"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/books/author/{author}": {
      "get": {
        "tags": ["Books"],
        "summary": "Get books by author",
        "description": "Retrieves all books by a specific author (case-insensitive partial match)",
        "operationId": "getBooksByAuthor",
        "parameters": [
          {
            "name": "author",
            "in": "path",
            "required": true,
            "description": "Author name to search for",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully retrieved books",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/BookDto"
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/books/low-stock": {
      "get": {
        "tags": ["Books"],
        "summary": "Get low stock books",
        "description": "Retrieves books with stock below the specified threshold",
        "operationId": "getLowStockBooks",
        "parameters": [
          {
            "name": "threshold",
            "in": "query",
            "required": false,
            "description": "Stock threshold (default: 10)",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 10
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully retrieved low stock books",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/BookDto"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/books/{id}/stock": {
      "patch": {
        "tags": ["Books"],
        "summary": "Update book stock",
        "description": "Updates the stock quantity for a book",
        "operationId": "updateStock",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "ID of the book to update stock",
            "schema": {
              "type": "integer",
              "format": "int64",
              "minimum": 1
            }
          },
          {
            "name": "quantity",
            "in": "query",
            "required": true,
            "description": "Quantity to add (positive) or remove (negative)",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Stock updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BookDto"
                }
              }
            }
          },
          "400": {
            "description": "Invalid quantity or insufficient stock",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Book not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT authentication token. Include in Authorization header as 'Bearer {token}'"
      }
    },
    "schemas": {
      "BookDto": {
        "type": "object",
        "required": ["isbn", "title", "author", "price", "stock", "category"],
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64",
            "description": "Unique identifier for the book",
            "readOnly": true
          },
          "isbn": {
            "type": "string",
            "minLength": 10,
            "maxLength": 13,
            "description": "International Standard Book Number (ISBN)",
            "example": "9781234567890"
          },
          "title": {
            "type": "string",
            "maxLength": 255,
            "description": "Title of the book",
            "example": "The Great Gatsby"
          },
          "author": {
            "type": "string",
            "maxLength": 255,
            "description": "Author of the book",
            "example": "F. Scott Fitzgerald"
          },
          "description": {
            "type": "string",
            "maxLength": 1000,
            "description": "Detailed description of the book",
            "example": "A classic American novel set in the Jazz Age"
          },
          "price": {
            "type": "number",
            "format": "decimal",
            "minimum": 0.01,
            "description": "Price of the book",
            "example": 12.99
          },
          "stock": {
            "type": "integer",
            "minimum": 0,
            "description": "Current stock quantity available",
            "example": 50
          },
          "category": {
            "type": "string",
            "maxLength": 100,
            "description": "Category or genre of the book",
            "example": "Fiction"
          }
        }
      },
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "timestamp": {
            "type": "string",
            "format": "date-time",
            "description": "Timestamp when the error occurred",
            "example": "2026-01-05T21:13:00"
          },
          "status": {
            "type": "integer",
            "description": "HTTP status code",
            "example": 404
          },
          "error": {
            "type": "string",
            "description": "Short error type description",
            "example": "Not Found"
          },
          "message": {
            "type": "string",
            "description": "Detailed error message explaining what went wrong",
            "example": "Book with ID 999 not found"
          },
          "path": {
            "type": "string",
            "description": "The request path that caused the error",
            "example": "/api/v1/books/999"
          }
        }
      }
    }
  }
}